# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/controls/tag": {
            "get": {
                "description": "List all tags.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get"
                ],
                "summary": "List all tags.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.Tag"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a tag.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "create"
                ],
                "summary": "Create a tag.",
                "parameters": [
                    {
                        "description": "Tag data",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Tag"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Tag"
                        }
                    }
                }
            }
        },
        "/controls/tag-type": {
            "get": {
                "description": "List all tag types.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get"
                ],
                "summary": "List all tag types.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.TagType"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a tag type.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "create"
                ],
                "summary": "Create a tag type.",
                "parameters": [
                    {
                        "description": "Tag Type data",
                        "name": "tag_type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.TagType"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.TagType"
                        }
                    }
                }
            }
        },
        "/controls/tag-type/{id}": {
            "get": {
                "description": "Get a tag type by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get"
                ],
                "summary": "Get a tag type by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TagType"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a tag type.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "Update a tag type.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tag Type data",
                        "name": "tag_type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.TagType"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "description": "Delete a tag type.",
                "tags": [
                    "delete"
                ],
                "summary": "Delete a tag type.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
        "/controls/tag/{id}": {
            "get": {
                "description": "Get a tag by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get"
                ],
                "summary": "Get a tag by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Tag"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a tag.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "Update a tag.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tag data",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Tag"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "description": "Delete a tag.",
                "tags": [
                    "delete"
                ],
                "summary": "Delete a tag.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.Tag": {
            "type": "object",
            "required": [
                "name",
                "tagType"
            ],
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tagType": {
                    "type": "object",
                    "required": [
                        "id"
                    ],
                    "properties": {
                        "colour": {
                            "type": "string"
                        },
                        "id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "updateUser": {
                    "type": "string"
                }
            }
        },
        "api.TagType": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "colour": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "rank": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Tag"
                    }
                },
                "updateUser": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}